Add the 'load more' option for more than 9 product

+++ controllers/categoryCtrl.js
@@ -1,5 +1,5 @@
 const Category = require('../models/categoryModel')
-
+const Products = require('../models/productModel')
 const categoryCtrl = {
 
 getCategories : async (req,res) =>{
@@ -27,6 +27,8 @@ createCategory:async (req,res)=>{
 },
 deleteCategory: async (req,res)=>{
     try {
+        const products = await Products.findOne({category: req.params.id})
+        if(products) return res.status(400).json({msg: 'please delete related products also'})
         await Category.findByIdAndDelete(req.params.id)
         res.json({msg: "category deleted successfully"})
     } catch (err) {
